+++ packages/router-lite/src/route-definition.ts
@@ -1,12 +1,41 @@
-import { IContainer, IModule, Metadata, onResolve } from '@aurelia/kernel';
-import { CustomElementDefinition, CustomElement } from '@aurelia/runtime-html';
+import {
+  IContainer,
+  IModule,
+  Metadata,
+  onResolve,
+  Protocol,
+} from '@aurelia/kernel';
+import {
+  CustomElementDefinition,
+  CustomElement,
+} from '@aurelia/runtime-html';
 
-import { TypedNavigationInstruction, NavigationInstructionType, Params } from './instructions.js';
-import { RouteConfig, IChildRouteConfig, Routeable, RouteType, Route, IRedirectRouteConfig } from './route.js';
-import { IRouteContext } from './route-context.js';
-import { isPartialChildRouteConfig, isPartialRedirectRouteConfig } from './validation.js';
-import { ensureArrayOfStrings, ensureString } from './util.js';
+import {
+  TypedNavigationInstruction,
+  NavigationInstructionType,
+  Params,
+} from './instructions.js';
+import {
+  RouteConfig,
+  IChildRouteConfig,
+  Routeable,
+  RouteType,
+  Route,
+  IRedirectRouteConfig,
+} from './route.js';
+import {
+  IRouteContext,
+} from './route-context.js';
+import {
+  isPartialChildRouteConfig,
+  isPartialRedirectRouteConfig,
+} from './validation.js';
+import {
+  ensureArrayOfStrings,
+  ensureString,
+} from './util.js';
 
+export const defaultViewportName = 'default';
 export class RouteDefinition {
   public readonly hasExplicitPath: boolean;
   public readonly caseSensitive: boolean;
@@ -15,18 +44,20 @@ export class RouteDefinition {
   public readonly viewport: string;
   public readonly id: string;
   public readonly data: Params;
+  public readonly fallback: string | null;
 
   public constructor(
-    public readonly config: Omit<RouteConfig, 'saveTo'>,
+    public readonly config: RouteConfig,
     public readonly component: CustomElementDefinition | null,
   ) {
     this.hasExplicitPath = config.path !== null;
     this.caseSensitive = config.caseSensitive;
     this.path = ensureArrayOfStrings(config.path ?? component!.name);
     this.redirectTo = config.redirectTo ?? null;
-    this.viewport = config.viewport ?? 'default';
+    this.viewport = config.viewport ?? defaultViewportName;
     this.id = ensureString(config.id ?? this.path);
     this.data = config.data ?? {};
+    this.fallback = config.fallback ?? null;
   }
 
   public static resolve(routeable: Promise<IModule>, context: IRouteContext): RouteDefinition | Promise<RouteDefinition>;
@@ -36,32 +67,26 @@ export class RouteDefinition {
   public static resolve(routeable: Routeable, context: IRouteContext): RouteDefinition | Promise<RouteDefinition>;
   public static resolve(routeable: Routeable, context?: IRouteContext): RouteDefinition | Promise<RouteDefinition> {
     if (isPartialRedirectRouteConfig(routeable)) {
-      return new RouteDefinition(routeable as RouteConfig, null);
+      return new RouteDefinition(RouteConfig.create(routeable, null), null);
     }
 
     // Check if this component already has a `RouteDefinition` associated with it, where the `config` matches the `RouteConfig` that is currently associated with the type.
     // If a type is re-configured via `Route.configure`, that effectively invalidates any existing `RouteDefinition` and we re-create and associate it.
     // Note: RouteConfig is associated with Type, but RouteDefinition is associated with CustomElementDefinition.
     return onResolve(this.resolveCustomElementDefinition(routeable, context), def => {
+      const type = def.Type;
       const config = isPartialChildRouteConfig(routeable)
-        ? {
-          ...Route.getConfig(def.Type),
-          ...routeable
-        }
+        ? Route.isConfigured(type)
+          ? Route.getConfig(type).applyChildRouteConfig(routeable)
+          : RouteConfig.create(routeable, type)
         : Route.getConfig(def.Type);
 
-      if (!Metadata.hasOwn(Route.name, def)) {
-        const routeDefinition = new RouteDefinition(config, def);
-        Metadata.define(Route.name, routeDefinition, def);
-      } else {
-        let routeDefinition = Metadata.getOwn(Route.name, def) as RouteDefinition;
-        if (routeDefinition.config !== config) {
-          routeDefinition = new RouteDefinition(config, def);
-          Metadata.define(Route.name, routeDefinition, def);
-        }
+      let routeDefinition = $RouteDefinition.get(def);
+      if (routeDefinition === null) {
+        routeDefinition = new RouteDefinition(config, def);
+        $RouteDefinition.define(routeDefinition, def);
       }
-
-      return Metadata.getOwn(Route.name, def) as RouteDefinition;
+      return routeDefinition;
     });
   }
 
@@ -109,9 +134,34 @@ export class RouteDefinition {
   public toString(): string {
     const path = this.config.path === null ? 'null' : `'${this.config.path}'`;
     if (this.component !== null) {
-      return `RD(config.path:${path},c.name:'${this.component.name}')`;
+      return `RD(config.path:${path},c.name:'${this.component.name}',vp:'${this.viewport}')`;
     } else {
       return `RD(config.path:${path},redirectTo:'${this.redirectTo}')`;
     }
   }
 }
+
+export const $RouteDefinition = {
+  name: Protocol.resource.keyFor('route-definition'),
+  /**
+   * Returns `true` if the `def` has a route definition.
+   */
+  isDefined(def: CustomElementDefinition): boolean {
+    return Metadata.hasOwn($RouteDefinition.name, def);
+  },
+  /**
+   * Apply the specified configuration to the specified type, overwriting any existing configuration.
+   */
+  define(routeDefinition: RouteDefinition, customElementDefinition: CustomElementDefinition): void {
+    Metadata.define($RouteDefinition.name, routeDefinition, customElementDefinition);
+  },
+  /**
+   * Get the `RouteDefinition` associated with the `customElementDefinition`.
+   * Returns `null` if no route definition is associated with the given `customElementDefinition`.
+   */
+  get(customElementDefinition: CustomElementDefinition): RouteDefinition | null {
+    return $RouteDefinition.isDefined(customElementDefinition)
+      ? Metadata.getOwn($RouteDefinition.name, customElementDefinition) as RouteDefinition
+      : null;
+  },
+};